+++ internal/reconciliation/deployer_limiter_test.go
@@ -1,4 +1,4 @@
-package main
+package reconciliation
 
 import (
 	"context"
